optimized Sequence.PRIMES
The next offsets are now calculated in a more efficient way: As it looks for primes, it checks (line 204) whether each number is divisible by the highest prime factor in the next step (since, based on how it is iterating through numbers, it will already not be divisible by the other prime factors), and if it isn't, then it adds it to the list of next offsets. Then, when it is time to upgrade to the next step + offsets (lines 208-222) the next offsets are already calculated.

With these optimizations, it can find the first 10,000 primes (up to 104729) in ~300 milliseconds.