added Grunt task to automatically concatenate all scripts into one
If you run `grunt watch`, then it will listen for changes to anything in the scripts or tests directories, and if you change them it will update tests.js and/or scripts.js to reflect the changes. Repositories using this repository then just have to include one file (or two, if they want tests as well) to be able to use the code here.